extract BlogPostContent component
